Filter in HtmlAttributes from Pro
