+++ src/blocks/loop-item/components/BlockSettings.jsx
@@ -1,12 +1,12 @@
 import { __ } from '@wordpress/i18n';
 import { BaseControl, Notice } from '@wordpress/components';
 
+import { OpenPanel } from '@edge22/components';
+
 import {
 	ApplyFilters,
-	OpenPanel,
 	URLControls,
 	TagNameControl,
-	HtmlAttributes,
 } from '@components/index.js';
 import { useBlockStyles } from '@hooks/useBlockStyles';
 
@@ -83,13 +83,6 @@ export function BlockSettings( {
 						</Notice>
 					</BaseControl>
 				) }
-
-				<HtmlAttributes
-					items={ htmlAttributes }
-					onAdd={ ( value ) => setAttributes( { htmlAttributes: value } ) }
-					onRemove={ ( value ) => setAttributes( { htmlAttributes: value } ) }
-					onChange={ ( value ) => setAttributes( { htmlAttributes: value } ) }
-				/>
 			</OpenPanel>
 		</ApplyFilters>
 	);
