+++ src/blocks/query/components/BlockSettings.jsx
@@ -1,9 +1,9 @@
 import { __ } from '@wordpress/i18n';
 
+import { OpenPanel } from '@edge22/components';
+
 import {
 	ApplyFilters,
-	OpenPanel,
-	HtmlAttributes,
 } from '@components/index.js';
 
 import { QueryInspectorControls } from './QueryInspectorControls';
@@ -14,10 +14,6 @@ export function BlockSettings( {
 	attributes,
 	setAttributes,
 } ) {
-	const {
-		htmlAttributes,
-	} = attributes;
-
 	const panelProps = {
 		name,
 		attributes,
@@ -46,14 +42,7 @@ export function BlockSettings( {
 				{ ...panelProps }
 				title={ __( 'Settings', 'generateblocks' ) }
 				panelId="settings"
-			>
-				<HtmlAttributes
-					items={ htmlAttributes }
-					onAdd={ ( value ) => setAttributes( { htmlAttributes: value } ) }
-					onRemove={ ( value ) => setAttributes( { htmlAttributes: value } ) }
-					onChange={ ( value ) => setAttributes( { htmlAttributes: value } ) }
-				/>
-			</OpenPanel>
+			/>
 		</ApplyFilters>
 	);
 }
