+++ src/blocks/shape/components/BlockSettings.jsx
@@ -1,14 +1,14 @@
 import { __ } from '@wordpress/i18n';
 
+import { OpenPanel } from '@edge22/components';
+
 import {
 	ApplyFilters,
-	OpenPanel,
 	IconControl,
 	ColorPickerControls,
-	moreDesignOptions,
 	UnitControl,
-	HtmlAttributes,
 } from '@components/index.js';
+import { moreDesignOptions } from '@utils';
 import { useBlockStyles } from '@hooks/useBlockStyles';
 
 export const shapeColorControls = [
@@ -33,7 +33,6 @@ export function BlockSettings( {
 	setAttributes,
 } ) {
 	const {
-		htmlAttributes,
 		html,
 	} = attributes;
 
@@ -110,14 +109,7 @@ export function BlockSettings( {
 				title={ __( 'Settings', 'generateblocks' ) }
 				shouldRender={ '' === currentAtRule }
 				panelId="settings"
-			>
-				<HtmlAttributes
-					items={ htmlAttributes }
-					onAdd={ ( value ) => setAttributes( { htmlAttributes: value } ) }
-					onRemove={ ( value ) => setAttributes( { htmlAttributes: value } ) }
-					onChange={ ( value ) => setAttributes( { htmlAttributes: value } ) }
-				/>
-			</OpenPanel>
+			/>
 		</ApplyFilters>
 	);
 }
