--- src/components/html-attributes/editor.scss
@@ -1,18 +1,0 @@
-.gb-html-attributes {
-	display: grid;
-	grid-template-columns: 1fr;
-	gap: 5px;
-
-	&__item {
-		display: grid;
-		grid-template-columns: 3fr 4fr 1fr;
-		align-items: center;
-		gap: 5px;
-
-		.components-base-control {
-			&, &__field {
-				margin: 0;
-			}
-		}
-	}
-}
