--- src/components/html-attributes/index.js
@@ -1,3 +1,0 @@
-import { HtmlAttributes } from './HtmlAttributes.jsx';
-
-export { HtmlAttributes };
