--- src/components/open-panel/editor.scss
@@ -1,13 +1,0 @@
-.gb-open-panel {
-	&__title {
-		display: flex;
-		justify-content: space-between;
-		align-items: center;
-		gap: 1rem;
-		margin-bottom: 1.5em;
-
-		h2 {
-			margin: 0;
-		}
-	}
-}
