--- src/components/open-panel/index.js
@@ -1,3 +1,0 @@
-import { OpenPanel, moreDesignOptions } from './OpenPanel.jsx';
-
-export { OpenPanel, moreDesignOptions };
